% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the fatal occurences for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number for the state.}

\item{year}{A number.}
}
\description{
This function takes a state number and year, and plots the location
of each occurence during that year and on that state on a map using
the \link[maps:map]{maps::map} function.
}
\note{
The functions is stopped and an error is returned if the file
containing that year's data cannot be read, or if the state's
number is invalid.
}
\examples{
\dontrun{fars_map_state(1, 2013)}

}
