% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read \code{fars} data from file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the file containing the data.}
}
\value{
A tibble (\link[dplyr:tbl_df]{dplyr::tbl_df}) with the data read from the file.
}
\description{
This function takes a path to a file containing \code{fars} data, reads
the data from the file, and returns it as a tibble
(\link[dplyr:tbl_df]{dplyr::tbl_df}).
}
\note{
The function is stopped and errors out if the given path does
not exist or is not a file.
}
\examples{
\dontrun{fars_read("data/accident_2013.csv.bz2")}

}
