% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads \code{fars} and filters data for the given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years.}
}
\value{
Returns a list of tibbles (\link[dplyr:tbl_df]{dplyr::tbl_df}), one for each
year given to the function.
}
\description{
This function takes a list of years and tries to read data from
files concerning those years. The data read contains only the
\code{MONTH} and \code{year} columns from the original data files.
}
\note{
This function expects that each year's file is in the working
directory under the name \verb{accident_\{year\}.csv.bz2}. If a year's
data can not be read that year is ignored and a warning is
issued, but all other data is still read.
}
\examples{
\dontrun{fars_read_years(c(2013,2014,2015))}

}
\seealso{
\link{make_filename}, \link{fars_read}
}
