% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of fatal occurences per month/year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years.}
}
\value{
A tibble (\link[dplyr:tbl_df]{dplyr::tbl_df}) where each row corresponds to a
month of the year, and every column other than the first
corresponds to a year.
}
\description{
Takes a list of years, and returns a summary of the number of
occurences fatal occurences per month and year.
}
\note{
Years whose data file cannot be bound are ignored.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014,2015))}

}
\seealso{
\link{fars_read_years}
}
