% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year the data concerns.}
}
\value{
Returns a string of the form \verb{accident_\{year\}.csv.bz2}
where year is the parameter given to the function.
}
\description{
This function takes a year, and returns the name of a file
containing the data from that year.
}
\note{
Parameter \code{year} is converted to an integer with the
\link{as.integer} function.
}
\examples{
make_filename(2020)
make_filename(2020.10) # Note how the year is converted to an integer

}
